mqtt/mqtt_subscriber/src/requirements.txt
paho-mqtt==1.5.1
pyserial==3.5
six==1.16.0
